+++ acquire-core-libs/tests/unit-tests.cpp
@@ -70,10 +70,11 @@ extern "C"
     // device-properties
     int unit_test__storage__storage_property_string_check();
     int unit_test__storage__copy_string();
-    int unit_test__storage_properties_set_array_extents();
-    int unit_test__storage_properties_set_chunking_props();
-    int unit_test__storage_properties_default_append_dimension_is_t();
-    int unit_test__storage_properties_set_sharding_props();
+    int unit_test__storage_properties_dimension_count();
+    int unit_test__storage_properties_set_append_dimension();
+    int unit_test__storage_properties_insert_dimension();
+    int unit_test__storage_properties_remove_dimension();
+    int unit_test__storage_properties_get_dimension();
     int unit_test__device_state_as_string__is_defined_for_all();
     int unit_test__device_kind_as_string__is_defined_for_all();
     int unit_test__sample_type_as_string__is_defined_for_all();
@@ -98,10 +99,11 @@ main()
         CASE(unit_test__monotonic_clock_increases_monotonically),
         CASE(unit_test__storage__storage_property_string_check),
         CASE(unit_test__storage__copy_string),
-        CASE(unit_test__storage_properties_set_array_extents),
-        CASE(unit_test__storage_properties_set_chunking_props),
-        CASE(unit_test__storage_properties_default_append_dimension_is_t),
-        CASE(unit_test__storage_properties_set_sharding_props),
+        CASE(unit_test__storage_properties_dimension_count),
+        CASE(unit_test__storage_properties_set_append_dimension),
+        CASE(unit_test__storage_properties_insert_dimension),
+        CASE(unit_test__storage_properties_remove_dimension),
+        CASE(unit_test__storage_properties_get_dimension),
         CASE(unit_test__device_state_as_string__is_defined_for_all),
         CASE(unit_test__device_kind_as_string__is_defined_for_all),
         CASE(unit_test__sample_type_as_string__is_defined_for_all),