+++ acquire-driver-common/tests/devkit/storage-get-meta.cpp
@@ -72,9 +72,6 @@ main()
                 storage = containerof(device, struct Storage, device);
 
                 CHECK(Device_Ok == storage_get_meta(storage, &metadata));
-                CHECK(0 == metadata.array_extents.is_supported);
-                CHECK(0 == metadata.chunk_shape.is_supported);
-                CHECK(0 == metadata.shard_shape_chunks.is_supported);
                 CHECK(0 == metadata.multiscale.is_supported);
 
                 CHECK(Device_Ok == driver_close_device(device));